[SYCL][NFC] Detach library loading helpers from UR
Functions to dynamically load a library and query a symbol out of it
should not be attached to UR, because they are used for other libraries
as well.

Moved them from `detail::ur` into `detail` namespace, outlined into a
seprate header and removed declarations from public SYCL headers.
